+++ src/Libraries/Microsoft.Extensions.AI.OpenAI/OpenAIChatClient.cs
@@ -10,6 +10,7 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.Json;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Shared.Diagnostics;
@@ -23,7 +24,7 @@ using OpenAI.Chat;
 namespace Microsoft.Extensions.AI;
 
 /// <summary>Represents an <see cref="IChatClient"/> for an OpenAI <see cref="OpenAIClient"/> or <see cref="ChatClient"/>.</summary>
-internal sealed class OpenAIChatClient : IChatClient
+internal sealed partial class OpenAIChatClient : IChatClient
 {
     // These delegate instances are used to call the internal overloads of CompleteChatAsync and CompleteChatStreamingAsync that accept
     // a RequestOptions. These should be replaced once a better way to pass RequestOptions is available.
@@ -154,10 +155,11 @@ internal sealed class OpenAIChatClient : IChatClient
                 input.Role == OpenAIClientExtensions.ChatRoleDeveloper)
             {
                 var parts = ToOpenAIChatContent(input.Contents);
+                string? name = SanitizeAuthorName(input.AuthorName);
                 yield return
-                    input.Role == ChatRole.System ? new SystemChatMessage(parts) { ParticipantName = input.AuthorName } :
-                    input.Role == OpenAIClientExtensions.ChatRoleDeveloper ? new DeveloperChatMessage(parts) { ParticipantName = input.AuthorName } :
-                    new UserChatMessage(parts) { ParticipantName = input.AuthorName };
+                    input.Role == ChatRole.System ? new SystemChatMessage(parts) { ParticipantName = name } :
+                    input.Role == OpenAIClientExtensions.ChatRoleDeveloper ? new DeveloperChatMessage(parts) { ParticipantName = name } :
+                    new UserChatMessage(parts) { ParticipantName = name };
             }
             else if (input.Role == ChatRole.Tool)
             {
@@ -230,7 +232,7 @@ internal sealed class OpenAIChatClient : IChatClient
                         new(ChatMessageContentPart.CreateTextPart(string.Empty));
                 }
 
-                message.ParticipantName = input.AuthorName;
+                message.ParticipantName = SanitizeAuthorName(input.AuthorName);
                 message.Refusal = refusal;
 
                 yield return message;
@@ -565,7 +567,6 @@ internal sealed class OpenAIChatClient : IChatClient
         result.TopP ??= options.TopP;
         result.PresencePenalty ??= options.PresencePenalty;
         result.Temperature ??= options.Temperature;
-        result.AllowParallelToolCalls ??= options.AllowMultipleToolCalls;
         result.Seed ??= options.Seed;
 
         if (options.StopSequences is { Count: > 0 } stopSequences)
@@ -584,6 +585,11 @@ internal sealed class OpenAIChatClient : IChatClient
                 {
                     result.Tools.Add(ToOpenAIChatTool(af, options));
                 }
+            }
+
+            if (result.Tools.Count > 0)
+            {
+                result.AllowParallelToolCalls ??= options.AllowMultipleToolCalls;
             }
 
             if (result.ToolChoice is null && result.Tools.Count > 0)
@@ -746,6 +752,27 @@ internal sealed class OpenAIChatClient : IChatClient
             _ => new ChatFinishReason(s),
         };
 
+    /// <summary>Sanitizes the author name to be appropriate for including as an OpenAI participant name.</summary>
+    private static string? SanitizeAuthorName(string? name)
+    {
+        if (name is not null)
+        {
+            const int MaxLength = 64;
+
+            name = InvalidAuthorNameRegex().Replace(name, string.Empty);
+            if (name.Length == 0)
+            {
+                name = null;
+            }
+            else if (name.Length > MaxLength)
+            {
+                name = name.Substring(0, MaxLength);
+            }
+        }
+
+        return name;
+    }
+
     /// <summary>POCO representing function calling info. Used to concatenation information for a single function call from across multiple streaming updates.</summary>
     private sealed class FunctionCallInfo
     {
@@ -753,4 +780,13 @@ internal sealed class OpenAIChatClient : IChatClient
         public string? Name;
         public StringBuilder? Arguments;
     }
+
+    private const string InvalidAuthorNamePattern = @"[^a-zA-Z0-9_]+";
+#if NET
+    [GeneratedRegex(InvalidAuthorNamePattern)]
+    private static partial Regex InvalidAuthorNameRegex();
+#else
+    private static Regex InvalidAuthorNameRegex() => _invalidAuthorNameRegex;
+    private static readonly Regex _invalidAuthorNameRegex = new(InvalidAuthorNamePattern, RegexOptions.Compiled);
+#endif
 }